GDPAPI-1979:  Added checks for handling groups correctly.
One of the groups (security identifier "S-1-18-1") could not be accessed and be cast as an NT Group object.
This is well known security identifier - "AUTHENTICATION_AUTHORITY_ASSERTED_IDENTITY" (see https://msdn.microsoft.com/en-us/library/cc980032.aspx).
The fix is to ignore anything that can't be cast to an NTAccount object.